requirements.txt
loguru==0.7.2
python-dotenv==1.0.1
aiohttp-socks==0.8.4
aiohttp==3.9